requirements.txt
noiseprotocol
cryptography
tpm2-pytss